Renamed folder navbar to navigation
