+++ src/controllers/MemeInfoController.js
@@ -1,4 +1,5 @@
 const Category = require('../models/category');
+const Meme = require('../models/meme');
 
 const fetchCategories = async (req, res, next) => {
 
@@ -10,4 +11,19 @@ const fetchCategories = async (req, res, next) => {
 
 };
 
+const fetchSitesCount = async (req, res, next) => {
+
+    const category = req.params.category;
+
+    const sites = await Meme.count({
+        where: category ? {
+            category_name: category
+        } : undefined
+    });
+
+    res.json({sites: Math.ceil(sites/10)});
+
+};
+
+exports.fetchSitesCount = fetchSitesCount;
 exports.fetchCategories = fetchCategories;
